// src/types/tvLinksProps.ts
export interface tvLinksProps {
    id: number;
    imdb_id: string;
    freebase_mid: null | string;
    freebase_id: null | string;
    tvdb_id: number;
    tvrage_id: null | number;
    wikidata_id: string;
    facebook_id: null | string;
    instagram_id: null | string;
    twitter_id: null | string;
}